dashboard estatistico para requisições do sistema
